Added MultiEconomyWrapper Moved economy wrappers to /economy/wrappers/ package Improved Economy documentation Improved IdentityEconomy documentation Improved MultiEconomy documentation
