+++ src/main/java/net/milkbowl/vault/economy/Economy.java
@@ -22,7 +22,6 @@ import org.bukkit.OfflinePlayer;
 
 /**
  * The main economy API
- *
  */
 public interface Economy {
