+++ src/main/java/net/milkbowl/vault/economy/wrappers/IdentityEconomyWrapper.java
@@ -1,5 +1,7 @@
-package net.milkbowl.vault.economy;
+package net.milkbowl.vault.economy.wrappers;
 
+import net.milkbowl.vault.economy.Economy;
+import net.milkbowl.vault.economy.IdentityEconomy;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.ServicePriority;
 import org.bukkit.plugin.ServicesManager;
@@ -12,8 +14,8 @@ public class IdentityEconomyWrapper {
     }
 
     /**
-     * Will register both IdentityEconomy and legacy Economy to Vault
-     * @return true registered successfully, false already registered
+     * Will register IdentityEconomy and legacy Economy to Vault
+     * @return true if registered successfully, false if already registered
      */
     public boolean registerProviders(){
         ServicesManager manager = Bukkit.getServicesManager();
